+++ server/rpm/pbench-server.spec.j2
@@ -106,6 +106,7 @@ fi
 /%{installdir}/lib/pbench/common/constants.py
 /%{installdir}/lib/pbench/common/exceptions.py
 /%{installdir}/lib/pbench/common/logger.py
+/%{installdir}/lib/pbench/common/utils.py
 /%{installdir}/lib/pbench/__init__.py
 /%{installdir}/lib/pbench/server/__init__.py
 /%{installdir}/lib/pbench/server/indexer.py
